requirements.txt
fastapi
uvicorn[standard]
redis
sqlalchemy
pydantic[email]
playwright
prometheus-fastapi-instrumentator
python-multipart
requests
urllib3
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
email-validator
python-dotenv
